move last components out of app
